refactor(sensors): move `ChannelsSamplesZip` into `ariel-os-sensors`
