+++ src/ariel-os-macros/src/define_count_adjusted_sensor_enums.rs
@@ -205,47 +205,6 @@ pub fn define_count_adjusted_sensor_enums(_item: TokenStream) -> TokenStream {
         enum InnerReadingChannels {
             #(#reading_channels_variants),*
         }
-
-        // Introducing a custom iterator type is necessary for type erasure.
-        struct ChannelsSamplesZip {
-            reading_channels: ReadingChannels,
-            samples: InnerSamples,
-            i: usize,
-        }
-
-        impl ChannelsSamplesZip {
-            fn new(reading_channels: ReadingChannels, samples: InnerSamples) -> Self {
-                Self {
-                    reading_channels,
-                    samples,
-                    i: 0,
-                }
-            }
-        }
-
-        impl Iterator for ChannelsSamplesZip {
-            type Item = (ReadingChannel, Sample);
-
-            fn next(&mut self) -> Option<Self::Item> {
-                // This is functionally zipping samples with channels.
-                // TODO: it might be possible to write this more efficiently.
-                match (self.reading_channels.iter().nth(self.i), self.samples.iter().nth(self.i)) {
-                    (Some(reading_channel), Some(sample)) => {
-                        self.i += 1;
-                        Some((reading_channel, sample))
-                    }
-                    _ => None,
-                }
-            }
-        }
-
-        impl ExactSizeIterator for ChannelsSamplesZip {
-            fn len(&self) -> usize {
-                self.reading_channels.iter().len().min(self.samples.iter().len())
-            }
-        }
-
-        impl core::iter::FusedIterator for ChannelsSamplesZip {}
     };
 
     TokenStream::from(expanded)
